+++ tests/unit_tests/Commands/Remote/DummyCommand.php
@@ -3,6 +3,7 @@
 namespace Pantheon\Terminus\UnitTests\Commands\Remote;
 
 use Pantheon\Terminus\Commands\Remote\SSHBaseCommand;
+use Symfony\Component\Console\Input\InputInterface;
 
 /**
  * Class DummyCommand
@@ -20,5 +21,14 @@ class DummyCommand extends SSHBaseCommand
     {
         $this->prepareEnvironment($site_env_id);
         return $this->executeCommand($dummy_args);
+    }
+
+    /**
+     * @param InputInterface $input
+     * @return bool|null
+     */
+    public function useUseTty($input)
+    {
+        return $this->useTty($input);
     }
 }