+++ tests/unit_tests/Models/UpstreamTest.php
@@ -2,9 +2,13 @@
 
 namespace Pantheon\Terminus\UnitTests\Models;
 
-use Pantheon\Terminus\Models\Site;
 use Pantheon\Terminus\Models\Upstream;
 
+/**
+ * Class UpstreamTest
+ * Tests the Pantheon\Terminus\Models\Upstream class
+ * @package Pantheon\Terminus\UnitTests\Models
+ */
 class UpstreamTest extends \PHPUnit_Framework_TestCase
 {
     /**
@@ -14,8 +18,8 @@ class UpstreamTest extends \PHPUnit_Framework_TestCase
     {
         $data = [
             'id' => 'upstream id',
-            'longname' => 'upstream label',
-            'machinename' => 'upstream machine name',
+            'label' => 'upstream label',
+            'machine_name' => 'upstream machine name',
         ];
 
         $model = new Upstream((object)$data);
@@ -23,59 +27,19 @@ class UpstreamTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * Tests the Upstream::parseAttributes() function
-     */
-    public function testParseAttributes()
-    {
-        $data = [
-            'product_id' => 'upstream id',
-            'url' => 'repository.url',
-            'branch' => 'repo branch',
-            'key' => 'value',
-        ];
-        $model = new Upstream((object)['attributes' => (object)$data,]);
-        foreach ($data as $key => $value) {
-            $this->assertEquals($value, $model->get($key));
-        }
-        $this->assertEquals($data['product_id'], $model->id);
-    }
-
-    /**
      * Tests the Upstream::serialize() function
      */
     public function testSerialize()
     {
-        $expected = $data = [
-            'product_id' => 'upstream id',
-            'url' => 'repository.url',
-            'branch' => 'repo branch',
+        $data = $expected = [
+            'repository_url' => 'repository.url',
+            'label' => 'upstream label',
+            'machine_name' => 'upstream machine name',
+            'id' => 'id',
             'key' => 'value',
         ];
-        $expected['id'] = $data['product_id'];
+        $expected['organization'] = null;
         $model = new Upstream((object)$data);
         $this->assertEquals($expected, $model->serialize());
-
-        unset($expected['id']);
-        unset($expected['key']);
-        $model2 = new Upstream((object)$data);
-        $site = $this->getMockBuilder(Site::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-        $site->id = 'site id';
-        $model2->site = $site;
-        $this->assertEquals($expected, $model2->serialize());
-    }
-
-    /**
-     * Tests the Upstream::__toString() function
-     */
-    public function testToString()
-    {
-        $data = [
-            'id' => 'upstream id',
-            'url' => 'repository.url',
-        ];
-        $model = new Upstream((object)$data);
-        $this->assertEquals("{$data['id']}: {$data['url']}", (string)$model);
     }
 }